Create CVData and Input
In the assets folder,
create a data file and
inside it, create
a data structure that
would hold the CV
data.

Create a components
folder and in it,
create an Input file.
In the file, create and
export an InputField
component and have it
take in 5 props.
Have it take inputName
as a prop so that
the input name can
be used to update the
corresponding data
in the data structure.

If the textArea InputField
is true, return a textarea
else return an input element.
The textarea will be used
for the responsibilities
field in the
work experience section.