Set up CVForm
In components, create an
CVForm file and inside it
create a component with
the same name.

Import useState, InputField
and CVData. Set the state to
the imported data. Define
a handleChange function
that takes in an event
and an idToUpdate.
In the handleChange(),
deconstruct the event.target
object into the name
and value variables.

Pass in the current state
in setData so that it
maps through the current
state and for each object
it checks if the
object's id is the same
with the idToUpdate. If
it is, it creates a new
object and copies all the
enumerable properties
from the this object
into the newly created
object.
It then uses computed
property names to
overwrite overwrite
any existing property
in this object that
has the same key as
the evaluated name
variable. Else any object
whose id doesnt match the
idToUpdate gets returned
untouched to the
final array that the
map method passes
to setData.

This will ensure that `data`
state will be updated
onChange of any field.
This can now
save the respective
information in the respective
object in the `data` state array

Return JSX that has
calls InputField
with respective
props for each
field needed in
the Basic information
section of the form.
Use the basic-information
object (id is 0).
For the onChange prop,
call handleChange with
the event and the
basic-information object's
id. Wrap all these in a
form element and add
a submit button.