Set up CVDisplay
Create a CVDisplay
component and in it,
define and export
a CVDisplay
component that takes
in CVDataArray and
handleEdit props.

It uses the data from
CVDataArray to render
the data in different
HTML sections, in form
of a CV.
Add an edit button
and call the
handleEdit prop
onClick of the
button. This
button would be used
to edit the CV.

This page would render
the CV when the submit
button on the CVForm
is clicked.