Display CV onSubmit
Import CV display in
CVForm. In the
component, define a
new setIsSet state
variable, its updater
function and an
initial state of
false.

Define a handleSubmit
event handler that takes
in an event,
prevents the event's
default behavior
and setIsSet to be
true.

Define a handleEdit
event handler that
setIsSet
to false.

Return the CVDisplay
component with
the data state and
handleEdit as
the values of its
respective props if
isSent is true,

Attach the
handleSubmit event
handler to the
form's onSubmit event.

With this setup,
onSubmit, isSent
is set to true and
the CVDisplay
is rendered with
all details that
were filled in the
inputs. onClick
of the Edit button,
isSent is set to true
and the CVForm is
rendered
with all inputs
populated
with respective
values from
the data state.