Start styling CVForm
In the CVForm component,
replace the JSX fragments
in the return statement
with a section element
so that it
can be easily styled.

In the Input component,
change the input element
structure such that the
label is not wrapped
around it. This is so
that the label can
be styled with a
block display and
have the input
under it.

In src directory, create
CVForm.css file and style
the form-body, input, label
and heading. Import it
into main.jsx

Replace the default
styles in index.css
with basic resetting
styles.