Add more styles to form.
In CVForm.jsx, add a
section title className
to the all h3 that
contained the section
titles. Add a className
to the work experience
section.

In the CVForm.css add
more styles to the
form-body and the input.
Add styles to the sections
and the
section-title h3.

In index.css, style
button.