Apply more styles to form
In the assets directory,
add a fonts folder
and add a new font.
Import the font in
CVForm and apply it
on label.
Add a font family
to font body and
update some of
its styles.

Add styles to form.

Group the textarea
with the input so
that they can share
the same styles.

Add a border and
a border radius to
section to
show demarcation
between sections
and add padding.