Restructure input and form
In Input.jsx, wrap the
input, textarea and
their labels in divs.
These would enable
them to be styled
together with their
labels.

In CVForm.jsx, wrap
the fields in the
basic-information
section with a
section element
in twos.Also wrap the
to and from fields
of the work-experience
section in a section
element.
This would enable them
to be placed side by
side on larger screens.