Add media queries
Add a media query for
all screens that are
from than 600px and
larger.
Style .subsection
ywith a
flex display and a gap.
Style subsection's div
children with flex: 1
so that the to and from
fields in work-experience
can stretch

Add a media query for
tablet screens.
Clamp the width of the
form-body between
3 values.

Add a media query for
screens screens of
1000px and larger
(desktops).
Style form-body with
a maximum width.