Style CVDisplay
In CVDisplay.jsx,
change the JSX
fragments wrapping
the whole display
elements into
a div element
with a class
of container.

Move the
basic-information
section the
age and sex above
the add email and
number. Add the
address.
Remove its h3
containing its section
title.

Create a CVDisplay
css file and
style the .container
and
.basic-information