Import and add icons
In the assets directort,
create images/icons
directory and in it
add 3 icons.

In CVDisplay.jsx,
import the icons
and use them
respectively
as the
src of the
img elements
beside the
email, phone
and address
details'

Wrap these details
and their respective
img elements in a
.details-container div.

These icons will be
used on wider screens.