Style basic-information
Style basic-information
with a flex display,
direction and gap
so that there can be
space between
the names,
email-and-phone and
address.

Increase font-size
of h2.

Style .email-and-phone
with flex display,
justify content
and gap.

Set the display
of img to be none
so that the icons
are not visible on
mobile devices.
This is applying
mobile-first design