Restructure CVDisplay
Give the section
title a class name
and wrap everything
else in a div.

Add the location
and wrap it in
a div with the
date-of-graduation.

Wrap the school name,
the course,
school-name and
certificate-obtained
in another div.

This will enable
them to be
styled with flex
to achieve the
target display.