Restructure CVDisplay
In the educational
section, separate
location and
date of
graduation into
different divs
so that they can
display one
above each other.
Wrap the other
details in
the educational
section in divs.
Each to its own
div. It is better
that they are inside
divs.

Restructure
the work experience
section by wrapping
every detail in its
own div. Also group
the first two details
in a div and the
remainder three
in another div.
This is so that this
section can be styled
similar to the
education section.