Style work-experience
In CVDisplay.jsx,
wrap work details
in a .work-details
div so that they
can be styled like
.education-details

In CVDisplay.css,
group .work-experience,
.work-details and
.work-details > div
with other selectors
that have a flex
play and column
flex-direction.

Group .work-experience
with
.educational-qualifications
so that it can give its
children a
gap.

Group .work-details
with.education-details
so that it can be
give its children a gap
and padding.

Group .work-details > div
with .education-details > div
so that all its direct
children divs can give
their children gaps.

These are done to apply
the same styles of
.educational-qualifications
to .work-experience
and maintain uniformity.