Change CVDisplay structure
In CVDisplay.jsx, wrap all
the JSX except the button in
a div. This is so that with
the flex display and a
min-height of 100vh for the
.container, this div can
be given a flex-grow of 1.
This will make it to
expand and push the
button to the bottom
of the page (in case this
div's content is small)

Add a class to the button
as it will have
peculiar styles.