Style container > div
In CVDisplay.css,
group container > div
with the selectors
at the top of the
CSS so that it is
styled with a flex
display and flex
direction of
column.

Give container
a minimum height
of 100vh so that
always takes up
at least the full
height of the
viewport.

Give the
.container > div
a gap so there is
space bbetween its
children and a
flex-grow value
of 1 so that it
takes more space
and
push the button to
the bottom of the
viewport on short
pages.

Add some right
and left margin
for title of
section and
reduce the left
and right padding
of education-details
and
work-details

Reduce the width
of edit button
and center it
with
margin 0 auto.

In index.css, reduce
the padding and
border radius of
button.