Add media queries
Add media queries
for all screens
from 600px and
wider.
In it, center
container and
give it a
box-shadow.
Reveal the images
(icons), set a
width and
set a
filter: invert(100%).
This filter inverts
the icon from black
to white.
Style .details-container
(the container containing
the icon and the details)
with a flex display,
justify content center and
gap.
Style .details-container > div
with align-self center
so that the detail
aligns itself vertically
with its icon.

Also a media query
for tablets - for
screens from
600px and wider but
not wider than
1000. In it,
style content with
a width that clamps
it between 3 values.