Add more styles for wider screens
In the media query for
screens from 600px and
wider, remove the margin
from the section titles.
Add a padding left and
right to educational
and work details,
change its flex direction
to row and add a gap.
This spreads its childrebn
apart horizontally

Have the RHS of educational
and work details take more
space than their sibling
because they need more
space.

Center graduation date.

Style from and to
accordingly so that
its children are
stacked one on top
the other and
centered.

Center dash, remove
the span that
contains the "-"
and add "to"
after dash with
a pseudo-element

In the default styles
before the media
query, style
from-and-to
so that the dates
are aligned
horizontally. This
applies for
mobile screens.