Style CVForm and CVDisplay
In CVForm.css,
add a background color
to the form-container.
Add an accent background
color and border shadow
to form-content.
Set its text to be
white which is
a complementary
color. Change the
border color of the
sections to white.

In CVDisplay.css, set
the button's
background-color
and color to
the accent and
complementary
colors respectively.